Add Hadoop SequenceFile support for tensorflow Dataset
This fix adds Hadoop SequenceFile for tensorflow Dataset,
so that it is possible to process files stored in Hadoop
system directly.

This fix is a very preliminary and early implementation.
It supports `org.apache.hadoop.io.Text` only, and there
is no compression support.

Will work on expanding the support of other seralization
types, and compresion support in the follow up PR.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>